move authoMiddleware.js and validation.js to utils directory
